+++ src/main/java/domain/bank/BettingMoney.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.bank;
+
+import domain.bank.Money;
 
 /**
  * @author 우가
