+++ src/main/java/domain/bank/Money.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.bank;
 
 import java.util.Objects;
